Replace property by name lookup with virtual dispatch
EffectWindow fetched properties from the relevant toplevel through the
Qt property system.

QObject::property(char*) is a fairly slow string search, (it's a loop
not a hash lookup!). It's fetched multiple times for every window in
every paint of some effects (such as blur).
Hotspot showed it as 5% of the standard render pass (X11) with nothing
in kwin animating.

We can't easily do something akin to QQmlPropertyCache as toplevel
changes at runtime.

All properties relate to the relevant methods of the toplevel so we can
just map to the property.